fix(security): resolve static analysis findings

+++ compileflow-core/src/main/java/com/alibaba/compileflow/engine/core/runtime/cache/DefaultProcessRuntimeCache.java
@@ -285,7 +285,7 @@ public class DefaultProcessRuntimeCache implements ProcessRuntimeCache {
         }
     }
 
-    private int residentRuntimeCount() {
+    private long residentRuntimeCount() {
         Set<ProcessRuntimeIdentity> identities = new HashSet<>(runtimes.asMap().keySet());
         identities.addAll(retainedRuntimes.keySet());
         return identities.size();
